+++ src/data/brigades.ts
@@ -1,11 +1,16 @@
 import { CommanderDescriptor, baseCommander } from "./commanders";
 import { UnitDescriptor } from "./units";
 
+export interface BrigadeUnitDescriptor {
+    count: number;
+    unit: UnitDescriptor;
+};
+
 export interface BrigadeDescriptor {
     id: string;
     name: string;
     commander: CommanderDescriptor;
-    units: UnitDescriptor[];
+    units: BrigadeUnitDescriptor[];
 };
 
 export const baseBrigade : BrigadeDescriptor = {