+++ src/logic/calcBrigadeCost.ts
@@ -5,7 +5,7 @@ import { calcUnitCost } from "./calcUnitCost";
 export function calcBrigadeCost(brigade: BrigadeDescriptor) : number {
 
     const commanderCost = calcCommanderCost(brigade.commander);
-    const unitsCost = brigade.units.reduce((sum, unit) => sum + calcUnitCost(unit), 0);
+    const unitsCost = brigade.units.reduce((sum, unit) => sum + (unit.count * calcUnitCost(unit.unit)), 0);
 
     return commanderCost + unitsCost;
 };