+++ src/logic/prepareBrigadeData.ts
@@ -1,7 +1,6 @@
 import { v4 as uuid } from "uuid"
-import { BrigadeDescriptor, baseBrigade } from "../data/brigades";
+import { BrigadeDescriptor, BrigadeUnitDescriptor, baseBrigade } from "../data/brigades";
 import { CommanderDescriptor } from "../data/commanders";
-import { UnitDescriptor } from "../data/units";
 
 export function prepareBrigadeData(input: object, fallback: BrigadeDescriptor = { ...baseBrigade }) : BrigadeDescriptor {
 
@@ -12,7 +11,7 @@ export function prepareBrigadeData(input: object, fallback: BrigadeDescriptor = 
 
     if ('name' in input) brigade.name = String(input.name);
     if ('commander' in input && input.commander) brigade.commander = input.commander as CommanderDescriptor;
-    if ('units' in input) brigade.units = input.units as UnitDescriptor[];
+    if ('units' in input) brigade.units = input.units as BrigadeUnitDescriptor[];
 
     return brigade;
 };